moves fake* into own package
